Move Importer class to __init__
